move tests from test_large_array.py to test_large_vector.py
